+++ includes/classes/pools/mpos.php
@@ -49,7 +49,7 @@ class Pools_Mpos extends Pools_Abstract {
             $data['balance'] = number_format($poolData['getuserbalance']['confirmed'], 8);
             $data['unconfirmed_balance'] = number_format($poolData['getuserbalance']['unconfirmed'], 8);
 
-            $data['network_hashrate'] = formatHashrate($poolData['getpoolstatus']['nethashrate']);
+            $data['network_hashrate'] = formatHashrate($poolData['getpoolstatus']['nethashrate']/1000);
 
             $data['pool_hashrate'] = formatHashrate($poolData['getpoolstatus']['hashrate']);